refactor: Rename clitool folder to client
Plan is to designate the client folder as where the Streamlit app and CLI-specific code lives.
